+++ PEA/BruteForce.h
@@ -8,13 +8,13 @@ namespace Algorithm
 	{
 		BruteForce() {}
 
-		static std::pair<int, std::vector<unsigned int>> getMinimalCycleRecur(std::shared_ptr<Graph> graph, 
-			unsigned int vertex, int cost, std::vector<unsigned int> path, std::vector<bool> visited);
+		static std::pair<long long, std::vector<unsigned long long>> getMinimalCycleRecur(std::shared_ptr<Graph> graph, 
+			unsigned long long vertex, long long cost, std::vector<unsigned long long> path, std::vector<bool> visited);
 
 	public:
-		static unsigned int * getMinimalCycleRecur(std::shared_ptr<Graph> graph);
-		static unsigned int * getMinimalCycleSlow(std::shared_ptr<Graph> graph);
-		static unsigned int * getMinimalCycle(std::shared_ptr<Graph> graph);
-		static unsigned int * getMinimalCycleOpt(std::shared_ptr<Graph> graph);
+		static unsigned long long * getMinimalCycleRecur(std::shared_ptr<Graph> graph);
+		static unsigned long long * getMinimalCycleSlow(std::shared_ptr<Graph> graph);
+		static unsigned long long * getMinimalCycle(std::shared_ptr<Graph> graph);
+		static unsigned long long * getMinimalCycleOpt(std::shared_ptr<Graph> graph);
 	};
 }
